[JENKINS-75012] Pre-compile Handlebars templates
