+++ pom.xml
@@ -50,6 +50,8 @@
     <spotbugs.effort>Max</spotbugs.effort>
     <spotbugs.threshold>Low</spotbugs.threshold>
     <spotless.check.skip>false</spotless.check.skip>
+    <node.version>20.11.0</node.version>
+    <npm.version>10.2.4</npm.version>
   </properties>
 
   <dependencyManagement>
