+++ src/main/webapp/templates/helpers/JSON2string.js
@@ -1,0 +1,3 @@
+export default function (object) {
+    return JSON.stringify(object);
+}
