+++ src/main/webapp/templates/helpers/addHierarchy.js
@@ -1,0 +1,5 @@
+export default function (context, parentHierarchy, options) {
+    if (parentHierarchy == undefined)
+    parentHierarchy = 0;
+    context["hierarchyLevel"] = parentHierarchy + 1;
+}
