+++ src/main/webapp/templates/helpers/addName.js
@@ -1,0 +1,5 @@
+export default function (name) {
+    var modName = "";
+    modName = name.replace(/[^a-z\d/-]+/gi, "_");
+    return modName;
+}
