+++ src/main/webapp/templates/helpers/addspaces.js
@@ -1,0 +1,6 @@
+export default function (hierarchyLevel) {
+    var spaces = 1.5;
+
+    spaces = spaces * hierarchyLevel;
+    return new Handlebars.SafeString(spaces);
+}
