+++ src/main/webapp/templates/helpers/applystatus.js
@@ -1,0 +1,15 @@
+export default function (status) {
+    var statusClass = "no_status";
+    switch (status) {
+        case "FAILED":
+            statusClass = "failed";
+            break;
+        case "PASSED":
+            statusClass = "passed";
+            break;
+        case "SKIPPED":
+            statusClass = "skipped";
+            break;
+    }
+    return statusClass;
+}
