+++ src/main/webapp/templates/index.js
@@ -1,0 +1,9 @@
+import Handlebars from 'handlebars';
+import tableBody from './table-body.hbs';
+import worstTestsTableBody from './worst-tests-table-body.hbs';
+
+window.Handlebars = Handlebars;
+window.testResultAnalyzerTemplates = {
+    'table-body': tableBody,
+    'worst-tests-table-body': worstTestsTableBody
+};
