+++ src/main/webapp/templates/worst-tests-table-content.hbs
@@ -1,0 +1,5 @@
+<div class="worst-tests-table-row">
+    <div class="table-cell row-heading">{{k}}</div>
+    <div class="table-cell build-result">{{v.length}}</div>
+    <div class="table-cell build-result">{{{buildLinks v}}}</div>
+</div>
